Updated model and generated code to have publishers and subscribers to get processed images from image_processor to user input panel.  Also fixed small compile warning in vertical and radial actuator controllers.

+++ code/agse_rosmod_project/01-Software-Configuration/agse/src/agse_package/src/agse_package/radial_actuator_controller.cpp
@@ -78,7 +78,7 @@ bool radial_actuator_controller::radialPosCallback(agse_package::radialPos::Requ
     // Business Logic for radialPos_server Server providing radialPos Service
   if (req.update == true)
     {
-      ROS_INFO("GOT NEW RADIAL GOAL: %d",req.goal);
+      ROS_INFO("GOT NEW RADIAL GOAL: %d",(int)req.goal);
       ROS_INFO("CURRENT RADIUS: %d",radialCurrent);
       radialGoal = req.goal;
     }
